# packages/contract/openapi.yaml
openapi: 3.0.3
info:
  title: MeterBase API
  version: "1.0.0"
  description: >
    Usage metering & monetization infrastructure. Ingest usage events, define
    meters, query time-bucketed aggregates, compute cost from pricing models,
    and fire usage-threshold alerts via signed webhooks.

    This document is the source of truth for the v1 REST contract. JSON
    properties are camelCase; the implementation maps to snake_case at the DB
    boundary. All endpoints are scoped to the organization that owns the API key.
servers:
  - url: http://localhost:48888
    description: Local
  - url: https://api.meterbase.example
    description: Production (placeholder)
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/meters:
    post:
      tags: [Meters]
      summary: Create a meter
      operationId: createMeter
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MeterCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Meter' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      tags: [Meters]
      summary: List meters
      operationId: listMeters
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: A page of meters
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Meter' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/meters/{slug}:
    parameters:
      - { $ref: '#/components/parameters/Slug' }
    get:
      tags: [Meters]
      summary: Get a meter
      operationId: getMeter
      responses:
        '200':
          description: The meter
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Meter' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Meters]
      summary: Delete a meter
      operationId: deleteMeter
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/meters/{slug}/query:
    parameters:
      - { $ref: '#/components/parameters/Slug' }
    get:
      tags: [Meters]
      summary: Query time-bucketed aggregates for a meter
      operationId: queryMeter
      parameters:
        - { in: query, name: from, required: true, schema: { type: string, format: date-time } }
        - { in: query, name: to, required: true, schema: { type: string, format: date-time } }
        - in: query
          name: windowSize
          required: true
          schema: { $ref: '#/components/schemas/WindowSize' }
        - { in: query, name: subject, schema: { type: string }, description: Filter by subject }
        - { in: query, name: customerId, schema: { type: string }, description: Filter by customer }
        - in: query
          name: groupBy
          description: Group results by one or more dimension paths (repeatable).
          schema: { type: array, items: { type: string } }
          style: form
          explode: true
      responses:
        '200':
          description: Aggregated usage series
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueryResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "Pass the API key as a bearer token. Authorization: Bearer <api_key>."
  parameters:
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
      description: Opaque pagination cursor returned as `nextCursor`.
    Slug:
      in: path
      name: slug
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Aggregation:
      type: string
      enum: [SUM, COUNT, AVG, MIN, MAX, UNIQUE_COUNT]
    WindowSize:
      type: string
      enum: [MINUTE, HOUR, DAY, MONTH]
...
    MeterCreate:
      type: object
      required: [slug, eventType, aggregation]
      properties:
        slug: { type: string, example: api_requests }
        eventType: { type: string, example: api_request }
        aggregation: { $ref: '#/components/schemas/Aggregation' }
        valueProperty:
          type: string
          nullable: true
          description: JSON path into data to aggregate (e.g. "tokens"). Null for COUNT.
          example: tokens
        groupBy:
          type: array
          items: { type: string }
          default: []
          description: Dimension paths in data (e.g. ["model","region"]).
    Meter:
      allOf:
        - $ref: '#/components/schemas/MeterCreate'
        - type: object
          required: [id, createdAt]
          properties:
            id: { type: string }
            createdAt: { type: string, format: date-time }
    QueryResult:
      type: object
      required: [meter, windowSize, data]
      properties:
        meter: { type: string }
        windowSize: { $ref: '#/components/schemas/WindowSize' }
        data:
          type: array
          items:
            type: object
            required: [bucket, value]
            properties:
              bucket: { type: string, format: date-time }
              value: { type: number }
              groups:
                type: object
                additionalProperties: { type: string }
                description: Present only when groupBy was supplied.
...
    Page:
      type: object
      properties:
        nextCursor:
          type: string
          nullable: true
          description: Pass as `cursor` to fetch the next page. Null when no more.
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }